+++ src/app/(post)/result/[postId]/layout.tsx
@@ -2,7 +2,6 @@
 
 import { PropsWithChildren, Suspense } from 'react';
 
-import CommentForm from '@/src/app/(post)/result/[postId]/_components/CommentForm';
 import Spinner from '@/src/components/Spinner';
 
 import TopBar from '../../../../components/Topbar';
@@ -17,7 +16,6 @@ const PostDetailLayout = ({ children }: PropsWithChildren) => {
       </TopBar.Container>
       <Suspense fallback={<Spinner text='게시글 정보를 불러오는 중 입니다.' />}>
         {children}
-        <CommentForm />
       </Suspense>
     </>
   );
